backend/requirements.txt
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
passlib
marshmallow
python-dotenv
gunicorn
flask_cors
requests
python-dotenv==0.21.0